🔧 ToolMode・FLOOR_TYPES をモジュール化
- toolMode.ts 新規作成: ToolMode 型、FLOOR_TYPES 配列、floorTypeToColor 関数を抽出
- 5ファイルの import を更新: App.tsx, FloorCanvas.tsx, useCanvasDraw.ts, usePointerHandlers.ts, export.ts
- Toolbar.tsx（dead code）を削除
